requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.13
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==1.4.49
psycopg2-binary==2.9.9
jinja2==3.1.3
supabase==2.15.3

